Add responsiveness to project buttons
This includes adding an devpost id to the Tipmate's project button
"See Devpost". This has more text than the other buttons, therefore
needed to resize the font and reduce the letter spacing more than the
other buttons. More styles were applied to reduce the gap and the
letter spacing of the buttons when screen becomes smaller.