Add all travel images
This includes adding all Imgur links inside the travel section of the
html, writing js code to select the appropriate size of image
(from Imgur). Seeing as most displays nowadays are retina, I decided to
go with an image resolution of 2x the size of the image on the webpage.

Also add a shuffle function that shuffles all the travel images whenever
the page loads, therefore resulting in a different order of images
each time.