Change div elmts to span elmts in project cards
Previously, there was a div element inside of a label element. I found
out that this is invalid html. So instead, change the div card element
inside the label elements to spans, and change the css display
property of these cards to block, effectively making the span element
appear like a div element.