Change src of travel images based on screen size
Problem: when the website was being viewed on a large screen, the travel
images were taking very long to load. This was because the original size
of the images were being downloaded by the browser (quite large).

Solution: make it so the largest size resolution for travel images is
1024 in the image's longest dimension. This is done by adding a suffix of
'l' at the end of the Imgur direct link, used in the imgs srcs.